+++ NEWS.md
@@ -14,6 +14,8 @@
 - New `--error` option, to cause functions returning `Result` to be mutated to return the
   specified error.
 
+- New `--no-config` option, to disable reading `.cargo/mutants.toml`.
+
 ## 1.2.2
 
 Released 2023-04-01